Cancel button was corrected

+++ UISettingsConfigurator.cpp
@@ -24,40 +24,60 @@ int UISettingsConfigurator::Configure(const TCHAR* szCfgInString, TCHAR** pszCfg
 		}
 	}
 
-	DrvFtaeAlarm::SettingsInitializer::CreateModule(GetModuleHandle("Drv_Ftae_HdaAlarm.dll"), std::shared_ptr<DrvFtaeAlarm::UIDialogViewController>());
-	if (_settingsDataSource) {
-		_settingsDataSource->SaveSettingsString("OutputXML.xml");
-	}
+	int isOk = DrvFtaeAlarm::SettingsInitializer::CreateModule(GetModuleHandle("Drv_Ftae_HdaAlarm.dll"), std::shared_ptr<DrvFtaeAlarm::UIDialogViewController>());
 	
-	HANDLE hFile = CreateFile("OutputXML.xml", GENERIC_READ, 0, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
-	if (hFile != INVALID_HANDLE_VALUE)
-	{
-		DWORD fileSizeHight = 0;
-		DWORD fileSizeLow = GetFileSize(hFile, &fileSizeHight);
-		if (fileSizeLow == INVALID_FILE_SIZE && fileSizeHight == NULL)
-		{
-			CloseHandle(hFile);
-			hFile = INVALID_HANDLE_VALUE;
-			return iRes;
+	if (isOk) {
+		if (_settingsDataSource) {
+			_settingsDataSource->SaveSettingsString("OutputXML.xml");
 		}
-		char* outStr = new char[fileSizeLow + 1];
-		DWORD numRead = 0;
-		BOOL isRead = FALSE;
-		isRead = ReadFile(hFile, outStr, fileSizeLow, &numRead, NULL);
-		if (isRead == FALSE || numRead != fileSizeLow)
+		HANDLE hFile = CreateFile("OutputXML.xml", GENERIC_READ, 0, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
+		if (hFile != INVALID_HANDLE_VALUE)
 		{
+			DWORD fileSizeHight = 0;
+			DWORD fileSizeLow = GetFileSize(hFile, &fileSizeHight);
+			if (fileSizeLow == INVALID_FILE_SIZE && fileSizeHight == NULL)
+			{
+				CloseHandle(hFile);
+				hFile = INVALID_HANDLE_VALUE;
+				return iRes;
+			}
+			char* outStr = new char[fileSizeLow + 1];
+			DWORD numRead = 0;
+			BOOL isRead = FALSE;
+			isRead = ReadFile(hFile, outStr, fileSizeLow, &numRead, NULL);
+			if (isRead == FALSE || numRead != fileSizeLow)
+			{
+				CloseHandle(hFile);
+				hFile = INVALID_HANDLE_VALUE;
+				return ODS::ERR::FILE;
+			}
+			*pszCfgOutString = outStr;
 			CloseHandle(hFile);
+			DeleteFile("OutputXML.xml");
+			DeleteFile(FITERS_XML_FILE_NAME);
+			DeleteFile(SETTINGS_XML_FILE_NAME);
 			hFile = INVALID_HANDLE_VALUE;
-			return ODS::ERR::FILE;
+			iRes = ODS::ERR::OK;
 		}
-		*pszCfgOutString = outStr;
-		CloseHandle(hFile);
-		DeleteFile("OutputXML.xml");
+		else {
+			DeleteFile("OutputXML.xml");
+			DeleteFile(FITERS_XML_FILE_NAME);
+			DeleteFile(SETTINGS_XML_FILE_NAME);
+		}
+	}
+	else {
 		DeleteFile(FITERS_XML_FILE_NAME);
 		DeleteFile(SETTINGS_XML_FILE_NAME);
-		hFile = INVALID_HANDLE_VALUE;
-		iRes = ODS::ERR::OK;
+		if (szCfgInString != NULL)
+		{
+			size_t len = _tcslen(szCfgInString);
+			char* outStr = new char[len + 1];
+			_tcscpy_s(outStr, len + 1, szCfgInString);
+			*pszCfgOutString = outStr;
+			iRes = ODS::ERR::OK;
+		}
 	}
+	
 	return iRes;
 }
